+++ internal/core/domain.go
@@ -54,7 +54,7 @@ func (s *DomainService) InsertDomain(ctx context.Context, site string) error {
 }
 
 // ListDomain lists all domains.
-func (s *DomainService) ListDomain(ctx context.Context, offset, limit int32) ([]DomainModel, error) {
+func (s *DomainService) ListDomain(ctx context.Context, offset, limit int64) ([]DomainModel, error) {
 	domains, err := s.q.ListDomain(ctx, db.ListDomainParams{
 		Offset: offset,
 		Limit:  limit,
@@ -86,7 +86,7 @@ func (s *DomainService) ListDomain(ctx context.Context, offset, limit int32) ([]
 }
 
 // ListDomainHeroes lists all domains.
-func (s *DomainService) ListDomainHeroes(ctx context.Context, offset, limit int32) ([]DomainModel, error) {
+func (s *DomainService) ListDomainHeroes(ctx context.Context, offset, limit int64) ([]DomainModel, error) {
 	domains, err := s.q.ListDomainHeroes(ctx, db.ListDomainHeroesParams{
 		Offset: offset,
 		Limit:  limit,
@@ -118,7 +118,7 @@ func (s *DomainService) ListDomainHeroes(ctx context.Context, offset, limit int3
 }
 
 // CrawlDomain lists all domains available for crawling
-func (s *DomainService) CrawlDomain(ctx context.Context, offset, limit int32) ([]DomainModel, error) {
+func (s *DomainService) CrawlDomain(ctx context.Context, offset, limit int64) ([]DomainModel, error) {
 	domains, err := s.q.CrawlDomain(ctx, db.CrawlDomainParams{
 		Offset: offset,
 		Limit:  limit,
@@ -204,7 +204,7 @@ func (s *DomainService) DisableDomain(ctx context.Context, domain string) error 
 }
 
 // GetDomainsByName returns a list of domains by name.
-func (s *DomainService) GetDomainsByName(ctx context.Context, searchString string, offset, limit int32) ([]DomainModel, error) {
+func (s *DomainService) GetDomainsByName(ctx context.Context, searchString string, offset, limit int64) ([]DomainModel, error) {
 	domains, err := s.q.GetDomainsByName(ctx, db.GetDomainsByNameParams{
 		Column1: NullString(searchString),
 		Offset:  offset,
